refactor: table comp
